backend/requirements.txt
ephem-4.1.5
motor-3.3.2
fastapi-0.110.0
fastapi-limiter-0.1.6
fastapi-mail-1.4.1
PyJWT-2.8.0
pymongo-4.6.2